rename TranslationCell -> TranslationViewCell
