+++ src/components/QuranReader/TranslationView/TranslationViewCell.module.scss
