+++ src/components/QuranReader/TranslationView/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
-import TranslationViewCell from './TranslationCell';
 import styles from './TranslationView.module.scss';
+import TranslationViewCell from './TranslationViewCell';
 
 import { QuranReaderStyles } from 'src/redux/slices/QuranReader/styles';
 import Verse from 'types/Verse';
@@ -13,8 +13,8 @@ type TranslationViewProps = {
 
 const TranslationView = ({ verses }: TranslationViewProps) => (
   <div className={styles.container}>
-    {verses.map((verse) => (
-      <TranslationViewCell verse={verse} isHighlighted={false} />
+    {verses.map((verse, index) => (
+      <TranslationViewCell verse={verse} isHighlighted={index === 2} />
     ))}
   </div>
 );
